requirements.txt
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.7
django-crispy-forms==1.8.1
django-filter==2.2.0
django-import-export==2.0
django-oauth-toolkit==1.2.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.2
Pillow==6.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
uritemplate==3.0.0
urllib3==1.25.7
xlrd==1.2.0
XlsxWriter==1.2.6
xlwt==1.3.0
